+++ lib/src/widgets/adw/new/scaffold.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:libadwaita/src/controllers/flap_controller.dart';
 import 'package:libadwaita/src/widgets/widgets.dart';
+import 'package:libadwaita_core/libadwaita_core.dart';
 
 class AdwScaffold extends StatefulWidget {
   const AdwScaffold({
@@ -8,9 +9,18 @@ class AdwScaffold extends StatefulWidget {
     required this.body,
     this.flap,
     this.flapStyle,
+    this.flapOptions,
     this.flapController,
-    this.headerbar,
+    @Deprecated('headerbar is deprecated, use the properties separately')
+        AdwHeaderBar? Function(Widget?)? headerbar,
     this.viewSwitcher,
+    this.viewSwitcherConstraint,
+    this.headerBarStyle,
+    this.start,
+    this.title,
+    this.end,
+    required this.actions,
+    this.controls,
   }) : super(key: key);
 
   final Widget body;
@@ -20,10 +30,19 @@ class AdwScaffold extends StatefulWidget {
   final FlapController? flapController;
 
   final FlapStyle? flapStyle;
+  final FlapOptions? flapOptions;
 
-  final Widget? Function(Widget? viewSwitcher)? headerbar;
+  final HeaderBarStyle? headerBarStyle;
+
+  final List<Widget>? start;
+  final Widget? title;
+  final List<Widget>? end;
+
+  final AdwActions actions;
+  final AdwControls? controls;
 
   final Widget? viewSwitcher;
+  final double? viewSwitcherConstraint;
 
   @override
   _AdwScaffoldState createState() => _AdwScaffoldState();
@@ -40,20 +59,34 @@ class _AdwScaffoldState extends State<AdwScaffold> {
 
   @override
   Widget build(BuildContext context) {
-    final isMobile = MediaQuery.of(context).size.width <= 600;
-    final headerbar =
-        widget.headerbar?.call(!isMobile ? widget.viewSwitcher : null);
-    final flap = widget.flap != null
+    final isMobile = MediaQuery.of(context).size.width <=
+        (widget.viewSwitcherConstraint ?? 600);
+    final isFlapVisible = widget.flap != null;
+    final isViewSwitcherVisible = widget.viewSwitcher != null;
+
+    final flap = isFlapVisible
         ? SizedBox(
             width: 200,
-            child: Drawer(elevation: 25, child: widget.flap!(true)),
+            child: Drawer(
+              elevation: 25,
+              child: widget.flap!(true),
+            ),
           )
         : null;
 
     return SafeArea(
       child: Column(
         children: [
-          if (headerbar != null) headerbar,
+          AdwHeaderBar(
+            actions: widget.actions,
+            controls: widget.controls,
+            title: isViewSwitcherVisible && !isMobile
+                ? widget.viewSwitcher
+                : widget.title,
+            end: widget.end ?? [],
+            start: widget.start ?? [],
+            style: widget.headerBarStyle ?? const HeaderBarStyle(),
+          ),
           Expanded(
             child: Scaffold(
               drawerEnableOpenDragGesture: _flapController
@@ -65,17 +98,20 @@ class _AdwScaffoldState extends State<AdwScaffold> {
               onDrawerChanged: _flapController?.onDrawerChanged,
               drawer: flap,
               endDrawer: flap,
-              body: widget.flap != null
+              body: isFlapVisible
                   ? AdwFlap(
                       flap: widget.flap!(false),
                       controller: widget.flapController,
                       style: widget.flapStyle,
+                      options: widget.flapOptions,
                       child: widget.body,
                     )
                   : widget.body,
-              bottomNavigationBar: widget.viewSwitcher != null && isMobile
+              bottomNavigationBar: isViewSwitcherVisible && isMobile
                   ? SizedBox(
-                      height: 51,
+                      height: widget.headerBarStyle != null
+                          ? widget.headerBarStyle!.height
+                          : 51,
                       child: Column(
                         mainAxisSize: MainAxisSize.min,
                         children: [